Rename ShopsOverlay → PlacesOverlay
to match other usages
