+++ app/src/main/res/layout/fragment_overlay_places.xml
